# specs/001-finance-hub/contracts/api.yaml
openapi: 3.1.0
info:
  title: Finance Hub API
  version: 1.0.0
  description: Self-hosted financial management API
  contact:
    name: Finance Hub
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
...
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        documentId:
          $ref: '#/components/schemas/UUID'
        expiresAt:
          type: string
          format: date-time

    # Import
...
paths:
...
  /workspaces/{workspaceId}/documents:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    get:
      tags: [Documents]
      summary: List documents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [inbox, linked, archived]
      responses:
        '200':
          description: Document list
  /workspaces/{workspaceId}/documents/upload-url:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    post:
      tags: [Documents]
      summary: Get presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mimeType, size]
              properties:
                filename:
                  type: string
                mimeType:
                  type: string
                size:
                  type: integer
                  maximum: 20971520
      responses:
        '200':
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
  /workspaces/{workspaceId}/documents/{documentId}/link:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: documentId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    post:
      tags: [Documents]
      summary: Link document to transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId]
              properties:
                transactionId:
                  $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: Document linked

  # Import
...